Refactored helper methods
