+++ node-runner-cli/github/github.py
@@ -1,0 +1,17 @@
+import json
+
+import requests
+from utils.utils import Helpers
+
+
+def latest_release(repo_name="radixdlt/radixdlt"):
+    req = requests.Request('GET',
+                           f'https://api.github.com/repos/{repo_name}/releases/latest')
+
+    prepared = req.prepare()
+    prepared.headers['Content-Type'] = 'application/json'
+    prepared.headers['user-agent'] = 'radixnode-cli'
+    resp = Helpers.send_request(prepared, print_response=False)
+
+    json_response = json.loads(resp.content)
+    return json_response["tag_name"]
