+++ node-runner-cli/radixnode.py
@@ -1,15 +1,16 @@
+import getpass
+import os
+import os.path
 import sys
 from argparse import ArgumentParser
-import os
-import subprocess
-import os.path
 from pathlib import Path
+
 import requests
-import pprint, json
-from requests.auth import HTTPBasicAuth
 import urllib3
-import getpass
-from datetime import datetime
+from github.github import latest_release
+from requests.auth import HTTPBasicAuth
+from utils.utils import run_shell_command
+from utils.utils import Helpers
 from version import __version__
 
 urllib3.disable_warnings()
@@ -37,37 +38,7 @@ def cli_version():
     return __version__
 
 
-def latest_release(repo_name="radixdlt/radixdlt"):
-    user = Helpers.get_nginx_user()
-    req = requests.Request('GET',
-                           f'https://api.github.com/repos/{repo_name}/releases/latest')
-    prepared = req.prepare()
-    resp = Helpers.send_request(prepared)
-    return resp.content.tag_name
 
-
-def printCommand(cmd):
-    print('-----------------------------')
-    if type(cmd) is list:
-        print('Running command :' + ' '.join(cmd))
-        print('-----------------------------')
-    else:
-        print('Running command ' + cmd)
-        print('-----------------------------')
-
-
-def run_shell_command(cmd, env=None, shell=False, fail_on_error=True, quite=False):
-    printCommand(cmd)
-    if env:
-        result = subprocess.run(cmd, env=env, shell=shell)
-    else:
-        result = subprocess.run(cmd, shell=shell)
-    if fail_on_error and result.returncode != 0:
-        print("""
-            Command failed. Exiting...
-        """)
-        sys.exit()
-    return result
 
 
 class Base:
@@ -102,7 +73,7 @@ class Base:
             print(f"Node key file already exist at location {keyfile_path}")
             keystore_password = getpass.getpass("Enter the password of the existing keystore file 'validator.ks':")
         else:
-            ask_keystore_exists = input("Do you have keystore file named 'validator.ks' already from previous node Y/n")
+            ask_keystore_exists = input("Do you have keystore file named 'validator.ks' already from previous node Y/n:")
             if ask_keystore_exists == "Y":
                 print(
                     f"Copy the keystore file 'validator.ks' to the location {keyfile_path} and then rerun the command")
@@ -112,6 +83,7 @@ class Base:
                 Generating new keystore file. Don't forget to backup the key from location {keyfile_path}/validator.ks
                 """)
                 keystore_password = getpass.getpass("Enter the password of the new file 'validator.ks':")
+                #TODO kegen image needs to be updated
                 run_shell_command(['docker', 'run', '--rm', '-v', keyfile_path + ':/keygen/key',
                                    'radixdlt/keygen:1.0-beta.31',
                                    '--keystore=/keygen/key/validator.ks',
@@ -409,72 +381,6 @@ class SystemD(Base):
         run_shell_command('sudo systemctl restart radixdlt-node.service', shell=True)
 
 
-class Helpers:
-    @staticmethod
-    def is_json(myjson):
-        try:
-            json_object = json.loads(myjson)
-        except ValueError as e:
-            return False
-        return True
-
-    @staticmethod
-    def pretty_print_request(req):
-        print('{}\n{}\r\n{}\r\n\r\n{}\n{}'.format(
-            '-----------Request-----------',
-            req.method + ' ' + req.url,
-            '\r\n'.join('{}: {}'.format(k, v) for k, v in req.headers.items()),
-            req.body,
-            '-------------------------------'
-        ))
-
-    @staticmethod
-    def send_request(prepared, print_request=False, print_response=True):
-        if print_request:
-            Helpers.pretty_print_request(prepared)
-        s = requests.Session()
-        resp = s.send(prepared, verify=False)
-        if Helpers.is_json(resp.content):
-            print(json.dumps(resp.json()))
-        else:
-            print(resp.content)
-        return resp
-
-    @staticmethod
-    def get_nginx_user():
-        nginx_admin_password = 'NGINX_ADMIN_PASSWORD'
-        if os.environ.get('%s' % nginx_admin_password) is None:
-            print("""
-            ------
-            NGINX_ADMIN_PASSWORD is missing !
-            Setup NGINX_ADMIN_PASSWORD environment variable using below command . Replace the string 'nginx_password_of_your_choice' with your password
-
-            echo 'export NGINX_ADMIN_PASSWORD="nginx_password_of_your_choice"' >> ~/.bashrc
-            """)
-            sys.exit()
-        else:
-            return dict({
-                "name": "admin",
-                "password": os.environ.get("%s" % nginx_admin_password)
-            })
-
-    @staticmethod
-    def docker_compose_down(composefile, remove_volumes):
-        command = ['docker-compose', '-f', composefile, 'down']
-        if remove_volumes:
-            command.append('-v')
-        run_shell_command(command)
-
-    @staticmethod
-    def get_public_ip():
-        return requests.get('https://api.ipify.org').text
-
-    @staticmethod
-    def get_current_date_time():
-        now = datetime.now()
-        return now.strftime("%Y_%m_%d_%H_%M")
-
-
 @subcommand([])
 def version(args):
     print(f"Cli - Version : {cli_version()}")
@@ -482,7 +388,7 @@ def version(args):
 
 @subcommand([
 
-    argument("-r", "--release", required=True,
+    argument("-r", "--release",
              help="Version of node software to install such as 1.0-beta.34",
              action="store"),
     argument("-n", "--nodetype", required=True, default="fullnode", help="Type of node fullnode or archivenode",
@@ -491,8 +397,12 @@ def version(args):
     argument("-u", "--update", help="Update the node to new version of composefile", action="store_false"),
 ])
 def setup_docker(args):
-    composefileurl = f"https://github.com/radixdlt/radixdlt/releases/download/{args.release}/radix-{args.nodetype}-compose.yml"
-    continue_setup = input(f"""Going to setup node type {args.nodetype} for version {args.release}
+    if not args.release:
+        release = latest_release()
+    else:
+        release = args.release
+    composefileurl = f"https://github.com/radixdlt/radixdlt/releases/download/{release}/radix-{args.nodetype}-compose.yml"
+    continue_setup = input(f"""Going to setup node type {args.nodetype} for version {release}
             From location {composefileurl}. Do you want to continue Y/n:
         """)
 
@@ -501,7 +411,7 @@ def setup_docker(args):
         sys.exit()
 
     Docker.fetchComposeFile(composefileurl)
-    keystore_password = Base.generatekey(keyfile_path=str(Path(__file__).parent.absolute()))
+    keystore_password = Base.generatekey(keyfile_path=os.getcwd())
     Base.fetch_universe_json(args.trustednode)
 
     compose_file_name = composefileurl.rsplit('/', 1)[-1]
@@ -608,7 +518,7 @@ def stop_systemd(args):
     argument("-t", "--trustednode", required=True, help="Trusted node on radix network", action="store")
 ])
 def start_docker(args):
-    keystore_password = Base.generatekey(keyfile_path=str(Path(__file__).parent.absolute()))
+    keystore_password = Base.generatekey(keyfile_path=os.getcwd())
     Docker.run_docker_compose_up(keystore_password, args.composefile, args.trustednode)
